refactor(websocket): 移除无用的ping函数和冗余的fin变量判断
1.  删除了未被调用的send_ping函数
2.  移除了process_websocket_frame和client_thread中未使用的fin变量定义
3.  修复base64解码时的类型转换警告

+++ utils/base64.c
@@ -129,7 +129,7 @@ static int __encodeData(void* struct_this, const void* d, size_t l, int isFinal)
     return __encode(struct_this);
 }
 
-#define SetTemp(_i, _input, _output) _output = 0xFF; for(_i=0; _i<BASE64_CHLEN; _i++) if(BASE64_CHARS[_i]==_input) _output=_i
+#define SetTemp(_i, _input, _output) _output = 0xFF; for(_i=0; _i<(size_t)BASE64_CHLEN; _i++) if(BASE64_CHARS[_i]==_input) _output=_i
 
 static int __decode(void* struct_this) {
     size_t tmp;
